+++ v_4_9_2/cloudconfig_test.go
@@ -1,4 +1,4 @@
-package v_4_10_0
+package v_4_9_2
 
 import (
 	"encoding/base64"
