+++ v_4_9_2/filemanager.go
@@ -1,4 +1,4 @@
-package v_4_10_0
+package v_4_9_2
 
 import (
 	"bytes"
@@ -12,7 +12,7 @@ import (
 )
 
 const (
-	version  = "v_4_10_0"
+	version  = "v_4_9_2"
 	filesDir = "files"
 )
