+++ v_4_9_2/files/config/kubelet-master.yaml.tmpl
