+++ v_4_9_2/files/config/kubelet-worker.yaml.tmpl
