+++ v_4_9_2/files/k8s-resource/calico-all.yaml
@@ -344,9 +344,6 @@ spec:
             requests:
               cpu: 50m
               memory: 100Mi
-            limits:
-              cpu: 50m
-              memory: 100Mi
           volumeMounts:
             - mountPath: /host/opt/cni/bin
               name: cni-bin-dir
