+++ v_4_9_2/files/k8s-resource/kube-proxy-ds.yaml
