+++ v_4_9_2/files/kubeconfig/addons.yaml
