+++ v_4_9_2/files/kubeconfig/kubelet-master.yaml
