+++ v_4_9_2/render_asset_content.go
@@ -1,4 +1,4 @@
-package v_4_10_0
+package v_4_9_2
 
 import (
 	"bytes"
