+++ v_4_9_2/worker_template.go
@@ -1,4 +1,4 @@
-package v_4_10_0
+package v_4_9_2
 
 const WorkerTemplate = `---
 ignition:
